update: allows data.table model in TransitionClassification to be matched by id
`model_by_id` argument is added to the constructor method. This flag is to indicate whether the `model` object is meant to be matched by the id column of the entity object in `x` or not. It should be noted that this flag only matters if the `model` object is of type [data.table::data.table()] where it must contains a numeric column called `prob` or list columns of type numeric and character called `probs` and `choices`. The model object must have a column which its name matches with the id column of the entity object in `x`.